replaces nav dropdowns with nav links to history and saved pages
