+++ src/components/Header.js
@@ -2,83 +2,30 @@ import React from 'react';
 import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
-import NavDropdown from 'react-bootstrap/NavDropdown';
-import DropdownListItem from './DropdownListItem';
 
-const Header = () => {
-  const store = {
-    1: {
-      id: 1,
-      title: 'How to Make Milk',
-      author: 'James Cameron',
-      image:
-        'https://images.unsplash.com/photo-1541963463532-d68292c34b19?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxleHBsb3JlLWZlZWR8Mnx8fGVufDB8fHx8&w=1000&q=80',
-      description: `It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using 'Content here, content here', making it look like readable English. Many desktop publishing packages and web page editors now use Lorem Ipsum as their default model text, and a search for 'lorem ipsum' will uncover many web sites still in their infancy. Various versions have evolved over the years, sometimes by accident, sometimes on purpose (injected humour and the like).`,
-      begginerWords: 334,
-      intermediateWords: 1093,
-      advancedWords: 443,
-    },
-    2: {
-      id: 1,
-      title: 'How to Make Milk',
-      author: 'James Cameron',
-      image:
-        'https://images.unsplash.com/photo-1541963463532-d68292c34b19?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxleHBsb3JlLWZlZWR8Mnx8fGVufDB8fHx8&w=1000&q=80',
-      description: `It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using 'Content here, content here', making it look like readable English. Many desktop publishing packages and web page editors now use Lorem Ipsum as their default model text, and a search for 'lorem ipsum' will uncover many web sites still in their infancy. Various versions have evolved over the years, sometimes by accident, sometimes on purpose (injected humour and the like).`,
-      begginerWords: 334,
-      intermediateWords: 1093,
-      advancedWords: 443,
-    },
-    3: {
-      id: 1,
-      title: 'How to Make Milk',
-      author: 'James Cameron',
-      image:
-        'https://images.unsplash.com/photo-1541963463532-d68292c34b19?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxleHBsb3JlLWZlZWR8Mnx8fGVufDB8fHx8&w=1000&q=80',
-      description: `It is a long established fact that a reader will be distracted by the readable content of a page when looking at its layout. The point of using Lorem Ipsum is that it has a more-or-less normal distribution of letters, as opposed to using 'Content here, content here', making it look like readable English. Many desktop publishing packages and web page editors now use Lorem Ipsum as their default model text, and a search for 'lorem ipsum' will uncover many web sites still in their infancy. Various versions have evolved over the years, sometimes by accident, sometimes on purpose (injected humour and the like).`,
-      begginerWords: 334,
-      intermediateWords: 1093,
-      advancedWords: 443,
-    },
-  };
-
-  return (
-    <Navbar className="color-nav" expand="lg" variant="light">
-      <Container>
-        <Navbar.Toggle aria-controls="basic-navbar-nav" />
-        <Navbar.Collapse id="basic-navbar-nav">
-          <Nav className="ms-auto" alignRight>
-            <NavDropdown
-              align="end"
-              title={<strong>Saved</strong>}
-              id="dropdown-menu-align-end"
-            >
-              <NavDropdown.Item href="#action/3.1">
-                Action Action Action Action Action Action Action Action Action
-                Action Action Action Action
-              </NavDropdown.Item>
-              <NavDropdown.Item href="#action/3.2">
-                Another action
-              </NavDropdown.Item>
-              <NavDropdown.Item href="#action/3.3">Something</NavDropdown.Item>
-              <NavDropdown.Divider />
-              <NavDropdown.Item href="#action/3.4">
-                Separated link
-              </NavDropdown.Item>
-            </NavDropdown>
-            <NavDropdown
-              align="end"
-              title="History"
-              id="dropdown-menu-align-end"
-            >
-              {Object.keys(store).map((id) => (
-                <DropdownListItem article={store[id]} />
-              ))}
-            </NavDropdown>
-          </Nav>
-        </Navbar.Collapse>
-      </Container>
-    </Navbar>
-  );
-};
+const Header = () => (
+  <Navbar className="color-nav" expand="lg" variant="light">
+    <Container>
+      <Navbar.Brand href="/">
+        <img
+          alt=""
+          src="/logo.svg"
+          width="30"
+          height="30"
+          className="d-inline-block align-top"
+        />{' '}
+        News-Sifter
+      </Navbar.Brand>
+      <Navbar.Toggle aria-controls="basic-navbar-nav" />
+      <Navbar.Collapse id="basic-navbar-nav">
+        <Nav className="ms-auto" alignRight>
+          <Nav.Link href="/Saved">
+            <strong>Saved</strong>
+          </Nav.Link>
+          <Nav.Link href="/History">History</Nav.Link>
+        </Nav>
+      </Navbar.Collapse>
+    </Container>
+  </Navbar>
+);
 export default Header;
