+++ src/components/Saved.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const Saved = () => <div>Saved</div>;
+
+export default Saved;
